Refactor: LetterListView 하위에 작성되어있던 LetterDetailView 코드를 별도의 파일로 분리
